added protect and authorize middleware to authenticate

+++ controllers/courses.js
@@ -1,7 +1,7 @@
 const ErrorResponse = require(`../utility/errorResponse`)
 const Course = require(`../models/Course`)
 const Bootcamp = require('../models/Bootcamp')
-const asyncHandler = require('../middleware/asyncHandler')
+const asyncHandler = require('../middleware/async')
 
 // @desc: Get all courses
 // @route: Get /api/v1/courses
@@ -44,7 +44,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @route: Post /api/v1/bootcamps/:bootcampId/courses
 // @access: Private
 exports.createCourse = asyncHandler(async (req, res, next) => {
-  console.log('create course'.red)
+  console.log('create course'.red.bgYellow)
 
   req.body.bootcamp = req.params.bootcampId
   const bootcamp = await Bootcamp.findById(req.params.bootcampId)
@@ -56,7 +56,23 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
       ),
     )
   }
+  // Check if user is allowed to create course for this bootcamp
+  if (bootcamp.user.toString() != req.user.id && req.user.role != 'admin') {
+    return next(
+      new ErrorResponse(
+        `The user ${req.user.name} is not allowed to add course for this bootcamp ${bootcamp.name}`,
+        401,
+      ),
+    )
+  }
+
+  req.body.user = req.user.id
+
   const course = await Course.create(req.body)
+
+  console.log(
+    `The course with id: ${course._id} is created successfully`.red.bgYellow,
+  )
   res.status(201).json({
     success: true,
     data: course,
@@ -67,20 +83,33 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
 // @route: Put /api/v1/courses/:id
 // @access: Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  console.log('update course'.red)
-  const updatedCourse = await Course.findByIdAndUpdate(
-    req.params.id,
-    req.body,
-    { new: true, runValidators: true },
-  )
-  if (!updatedCourse) {
+  console.log('Update course'.red.bgYellow)
+  let courseToUpdate = await Course.findById(req.params.id)
+  if (!courseToUpdate) {
     return next(
       new ErrorResponse(`Course not found with id of ${req.params.id}`, 404),
     )
   }
+  if (courseToUpdate.user.toString() != req.user.id) {
+    return next(
+      new ErrorResponse(
+        `The user ${req.user.name} is not allowed to update this course ${courseToUpdate.title}`,
+        401,
+      ),
+    )
+  }
+
+  courseToUpdate = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  })
+  console.log(
+    `The course with id: ${courseToUpdate._id} is updated successfully`.red
+      .bgYellow,
+  )
   res.status(200).json({
     success: true,
-    data: updatedCourse,
+    data: courseToUpdate,
   })
 })
 
@@ -88,13 +117,33 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @route: Delete /api/v1/courses:id
 // @access: Private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
-  console.log('delete course'.red)
-  const courseToDelete = await Course.findByIdAndDelete(req.params.id)
+  console.log('Delete course'.red.bgYellow)
+  let courseToDelete = await Course.findById(req.params.id)
+
+  // Check if course exists
   if (!courseToDelete) {
     return next(
       new ErrorResponse(`Course not found with id of ${req.params.id}`, 404),
     )
   }
+  // Check if user is allowed
+  if (
+    courseToDelete.user.toString() != req.user.id &&
+    req.user.role != 'admin'
+  ) {
+    return next(
+      new ErrorResponse(
+        `The user ${req.user.name} is not allowed to delete this course ${courseToDelete.title}`,
+        403,
+      ),
+    )
+  }
+
+  courseToDelete = await Course.findByIdAndDelete(req.params.id)
+  console.log(
+    `The course with id: ${courseToDelete._id} is deleted successfully.`.red
+      .bgYellow,
+  )
   res.status(200).json({
     success: true,
     data: courseToDelete,
